+++ backend/lib/edgehog_web/schema/devices_types.ex
@@ -23,6 +23,222 @@ defmodule EdgehogWeb.Schema.DevicesTypes do
   use Absinthe.Relay.Schema.Notation, :modern
 
   alias EdgehogWeb.Resolvers
+  alias EdgehogWeb.Middleware
+
+  @desc """
+  Describes a set of filters to apply when fetching a list of devices.
+
+  When multiple filters are specified, they are applied in an AND fashion to \
+  further refine the results.
+  """
+  input_object :device_filter do
+    @desc "Whether to return devices connected or not to Astarte."
+    field :online, :boolean
+
+    @desc """
+    A string to match against the device ID. The match is case-insensitive \
+    and tests whether the string is included in the device ID.
+    """
+    field :device_id, :string
+
+    @desc """
+    A string to match against the part number of the device's system model.
+    The match is case-insensitive and tests whether the string is included in \
+    the part number of the device's system model.
+    """
+    field :system_model_part_number, :string
+
+    @desc """
+    A string to match against the handle of the device's system model.
+    The match is case-insensitive and tests whether the string is included in \
+    the handle of the device's system model.
+    """
+    field :system_model_handle, :string
+
+    @desc """
+    A string to match against the name of the device's system model.
+    The match is case-insensitive and tests whether the string is included in \
+    the name of the device's system model.
+    """
+    field :system_model_name, :string
+
+    @desc """
+    A string to match against the part number of the device's hardware type.
+    The match is case-insensitive and tests whether the string is included in \
+    the part number of the device's hardware type.
+    """
+    field :hardware_type_part_number, :string
+
+    @desc """
+    A string to match against the handle of the device's hardware type.
+    The match is case-insensitive and tests whether the string is included in \
+    the handle of the device's hardware type.
+    """
+    field :hardware_type_handle, :string
+
+    @desc """
+    A string to match against the name of the device's hardware type.
+    The match is case-insensitive and tests whether the string is included in \
+    the name of the device's hardware type.
+    """
+    field :hardware_type_name, :string
+
+    @desc """
+    A string to match against the tags of the device.
+    The match is case-insensitive and tests whether the string is included in \
+    one of the tags of the device.
+    """
+    field :tag, :string
+  end
+
+  @desc """
+  An input object for a device attribute.
+  """
+  input_object :device_attribute_input do
+    @desc "The namespace of the device attribute."
+    field :namespace, non_null(:device_attribute_namespace)
+
+    @desc "The key of the device attribute."
+    field :key, non_null(:string)
+
+    @desc "The type of the device attribute."
+    field :type, non_null(:variant_type)
+
+    @desc "The value of the device attribute."
+    field :value, non_null(:variant_value)
+  end
+
+  @desc "The possible namespace values for device attributes"
+  enum :device_attribute_namespace do
+    @desc "Custom attributes, user defined"
+    value :custom
+  end
+
+  object :device_attribute do
+    @desc "The namespace of the device attribute."
+    field :namespace, non_null(:device_attribute_namespace)
+
+    @desc "The key of the device attribute."
+    field :key, non_null(:string)
+
+    @desc "The type of the device attribute."
+    field :type, non_null(:variant_type) do
+      resolve &Resolvers.Devices.extract_attribute_type/3
+    end
+
+    @desc "The value of the device attribute."
+    field :value, non_null(:variant_value) do
+      resolve &Resolvers.Devices.extract_attribute_value/3
+    end
+  end
+
+  @desc """
+  Denotes a device instance that connects and exchanges data.
+
+  Each Device is associated to a specific SystemModel, which in turn is \
+  associated to a specific HardwareType.
+  A Device also exposes info about its connection status and some sets of \
+  data read by its operating system.
+  """
+  node object(:device) do
+    @desc "The display name of the device."
+    field :name, non_null(:string)
+
+    @desc "The device ID used to connect to the Astarte cluster."
+    field :device_id, non_null(:string)
+
+    @desc "Tells whether the device is connected or not to Astarte."
+    field :online, non_null(:boolean)
+
+    @desc "The date at which the device last connected to Astarte."
+    field :last_connection, :datetime
+
+    @desc "The date at which the device last disconnected from Astarte."
+    field :last_disconnection, :datetime
+
+    @desc "The system model of the device."
+    field :system_model, :system_model
+
+    @desc "The tags of the device"
+    field :tags, non_null(list_of(non_null(:string))) do
+      resolve &Resolvers.Devices.extract_device_tags/3
+    end
+
+    @desc "The custom attributes of the device. These attributes are user editable."
+    field :custom_attributes, non_null(list_of(non_null(:device_attribute)))
+
+    @desc "List of capabilities supported by the device."
+    field :capabilities, non_null(list_of(non_null(:device_capability))) do
+      resolve &Resolvers.Capabilities.list_device_capabilities/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "Info read from the device's hardware."
+    field :hardware_info, :hardware_info do
+      resolve &Resolvers.Astarte.fetch_hardware_info/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "The estimated location of the device."
+    field :location, :device_location do
+      resolve &Resolvers.Geolocation.fetch_device_location/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "The current usage of the storage units of the device."
+    field :storage_usage, list_of(non_null(:storage_unit)) do
+      resolve &Resolvers.Astarte.fetch_storage_usage/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "The current status of the operating system of the device."
+    field :system_status, :system_status do
+      resolve &Resolvers.Astarte.fetch_system_status/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "The list of WiFi Access Points found by the device."
+    field :wifi_scan_results, list_of(non_null(:wifi_scan_result)) do
+      resolve &Resolvers.Astarte.fetch_wifi_scan_results/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "The status of the battery slots of the device."
+    field :battery_status, list_of(non_null(:battery_slot)) do
+      resolve &Resolvers.Astarte.fetch_battery_status/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "Information about the operating system's base image for the device."
+    field :base_image, :base_image do
+      resolve &Resolvers.Astarte.fetch_base_image/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "Information about the operating system of the device."
+    field :os_info, :os_info do
+      resolve &Resolvers.Astarte.fetch_os_info/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "The existing OTA operations for this device"
+    field :ota_operations, non_null(list_of(non_null(:ota_operation))) do
+      # TODO: this causes an N+1 if used on the device list, we should use dataloader instead
+      resolve &Resolvers.OSManagement.ota_operations_for_device/3
+    end
+
+    @desc "The status of cellular connection of the device."
+    field :cellular_connection, list_of(non_null(:modem)) do
+      resolve &Resolvers.Astarte.fetch_cellular_connection/3
+      middleware Middleware.ErrorHandler
+    end
+
+    @desc "Information about the Edgehog runtime running on the device."
+    field :runtime_info, :runtime_info do
+      resolve &Resolvers.Astarte.fetch_runtime_info/3
+      middleware Middleware.ErrorHandler
+    end
+  end
 
   @desc """
   Denotes a type of hardware that devices can have.
@@ -83,6 +299,22 @@ defmodule EdgehogWeb.Schema.DevicesTypes do
   end
 
   object :devices_queries do
+    @desc "Fetches the list of all devices."
+    field :devices, non_null(list_of(non_null(:device))) do
+      @desc "An optional set of filters to apply when fetching the devices."
+      arg :filter, :device_filter
+      resolve &Resolvers.Devices.list_devices/3
+    end
+
+    @desc "Fetches a single device."
+    field :device, :device do
+      @desc "The ID of the device."
+      arg :id, non_null(:id)
+
+      middleware Absinthe.Relay.Node.ParseIDs, id: :device
+      resolve &Resolvers.Devices.find_device/2
+    end
+
     @desc "Fetches the list of all hardware types."
     field :hardware_types, non_null(list_of(non_null(:hardware_type))) do
       resolve &Resolvers.Devices.list_hardware_types/3
@@ -113,6 +345,31 @@ defmodule EdgehogWeb.Schema.DevicesTypes do
   end
 
   object :devices_mutations do
+    @desc "Updates a device."
+    payload field :update_device do
+      input do
+        @desc "The GraphQL ID (not the Astarte Device ID) of the device to be updated."
+        field :device_id, non_null(:id)
+
+        @desc "The display name of the device."
+        field :name, :string
+
+        @desc "The tags of the device. These replace all the current tags."
+        field :tags, list_of(non_null(:string))
+
+        @desc "The custom attributes of the device. These replace all the current custom attributes."
+        field :custom_attributes, list_of(non_null(:device_attribute_input))
+      end
+
+      output do
+        @desc "The updated device."
+        field :device, non_null(:device)
+      end
+
+      middleware Absinthe.Relay.Node.ParseIDs, device_id: :device
+      resolve &Resolvers.Devices.update_device/2
+    end
+
     @desc "Creates a new hardware type."
     payload field :create_hardware_type do
       input do
