+++ backend/test/edgehog/astarte_test.exs
@@ -23,7 +23,6 @@ defmodule Edgehog.AstarteTest do
   use Edgehog.AstarteMockCase
 
   alias Edgehog.Astarte
-  alias Edgehog.Devices
 
   describe "clusters" do
     alias Edgehog.Astarte.Cluster
@@ -143,7 +142,6 @@ defmodule Edgehog.AstarteTest do
 
   describe "devices" do
     alias Edgehog.Astarte.Device
-    alias Edgehog.Astarte.InterfaceVersion
 
     import Edgehog.AstarteFixtures
     import Edgehog.DevicesFixtures
@@ -156,307 +154,8 @@ defmodule Edgehog.AstarteTest do
 
     @invalid_attrs %{device_id: nil, name: nil}
 
-    test "list_devices/0 returns all devices", %{realm: realm} do
-      device = device_fixture(realm)
-      assert Astarte.list_devices() == [device]
-    end
-
-    test "list_devices/1 filters with online", %{realm: realm} do
-      device_1 = device_fixture(realm, device_id: "7mcE8JeZQkSzjLyYuh5N9A", online: true)
-      _device_2 = device_fixture(realm, device_id: "nWwr7SZiR8CgZN_uKHsAJg", online: false)
-      filters = %{online: true}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with device_id", %{realm: realm} do
-      device_1 = device_fixture(realm, device_id: "7mcE8JeZQkSzjLyYuh5N9A")
-      _device_2 = device_fixture(realm, device_id: "nWwr7SZiR8CgZN_uKHsAJg")
-      filters = %{device_id: "7mc"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with system_model_part_number", %{realm: realm} do
-      hardware_type = hardware_type_fixture()
-
-      system_model_part_number_1 = "XYZ/1234"
-
-      _system_model_1 =
-        system_model_fixture(hardware_type,
-          name: "Foo",
-          handle: "foo",
-          part_numbers: [system_model_part_number_1]
-        )
-
-      system_model_part_number_2 = "ABC/0987"
-
-      _system_model_2 =
-        system_model_fixture(hardware_type,
-          name: "Bar",
-          handle: "bar",
-          part_numbers: [system_model_part_number_2]
-        )
-
-      device_1 =
-        device_fixture(realm,
-          device_id: "7mcE8JeZQkSzjLyYuh5N9A",
-          part_number: system_model_part_number_1
-        )
-
-      _device_2 =
-        device_fixture(realm,
-          device_id: "nWwr7SZiR8CgZN_uKHsAJg",
-          part_number: system_model_part_number_2
-        )
-
-      filters = %{system_model_part_number: "XYZ"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with system_model_name", %{realm: realm} do
-      hardware_type = hardware_type_fixture()
-
-      system_model_part_number_1 = "XYZ/1234"
-
-      _system_model_1 =
-        system_model_fixture(hardware_type,
-          name: "Foo",
-          handle: "foo",
-          part_numbers: [system_model_part_number_1]
-        )
-
-      system_model_part_number_2 = "ABC/0987"
-
-      _system_model_2 =
-        system_model_fixture(hardware_type,
-          name: "Bar",
-          handle: "bar",
-          part_numbers: [system_model_part_number_2]
-        )
-
-      device_1 =
-        device_fixture(realm,
-          device_id: "7mcE8JeZQkSzjLyYuh5N9A",
-          part_number: system_model_part_number_1
-        )
-
-      _device_2 =
-        device_fixture(realm,
-          device_id: "nWwr7SZiR8CgZN_uKHsAJg",
-          part_number: system_model_part_number_2
-        )
-
-      filters = %{system_model_name: "oo"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with system_model_handle", %{realm: realm} do
-      hardware_type = hardware_type_fixture()
-
-      system_model_part_number_1 = "XYZ/1234"
-
-      _system_model_1 =
-        system_model_fixture(hardware_type,
-          name: "Foo",
-          handle: "foo",
-          part_numbers: [system_model_part_number_1]
-        )
-
-      system_model_part_number_2 = "ABC/0987"
-
-      _system_model_2 =
-        system_model_fixture(hardware_type,
-          name: "Bar",
-          handle: "bar",
-          part_numbers: [system_model_part_number_2]
-        )
-
-      device_1 =
-        device_fixture(realm,
-          device_id: "7mcE8JeZQkSzjLyYuh5N9A",
-          part_number: system_model_part_number_1
-        )
-
-      _device_2 =
-        device_fixture(realm,
-          device_id: "nWwr7SZiR8CgZN_uKHsAJg",
-          part_number: system_model_part_number_2
-        )
-
-      filters = %{system_model_name: "fo"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with hardware_type_part_number", %{realm: realm} do
-      hardware_type_1 =
-        hardware_type_fixture(name: "HW1", handle: "hw1", part_numbers: ["AAA-BBB"])
-
-      system_model_part_number_1 = "XYZ/1234"
-
-      _system_model_1 =
-        system_model_fixture(hardware_type_1,
-          name: "Foo",
-          handle: "foo",
-          part_numbers: [system_model_part_number_1]
-        )
-
-      hardware_type_2 =
-        hardware_type_fixture(name: "HW2", handle: "hw2", part_numbers: ["CCC-DDD"])
-
-      system_model_part_number_2 = "ABC/0987"
-
-      _system_model_2 =
-        system_model_fixture(hardware_type_2,
-          name: "Bar",
-          handle: "bar",
-          part_numbers: [system_model_part_number_2]
-        )
-
-      device_1 =
-        device_fixture(realm,
-          device_id: "7mcE8JeZQkSzjLyYuh5N9A",
-          part_number: system_model_part_number_1
-        )
-
-      _device_2 =
-        device_fixture(realm,
-          device_id: "nWwr7SZiR8CgZN_uKHsAJg",
-          part_number: system_model_part_number_2
-        )
-
-      filters = %{hardware_type_part_number: "AAA"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with hardware_type_name", %{realm: realm} do
-      hardware_type_1 =
-        hardware_type_fixture(name: "HW1", handle: "hw1", part_numbers: ["AAA-BBB"])
-
-      system_model_part_number_1 = "XYZ/1234"
-
-      _system_model_1 =
-        system_model_fixture(hardware_type_1,
-          name: "Foo",
-          handle: "foo",
-          part_numbers: [system_model_part_number_1]
-        )
-
-      hardware_type_2 =
-        hardware_type_fixture(name: "HW2", handle: "hw2", part_numbers: ["CCC-DDD"])
-
-      system_model_part_number_2 = "ABC/0987"
-
-      _system_model_2 =
-        system_model_fixture(hardware_type_2,
-          name: "Bar",
-          handle: "bar",
-          part_numbers: [system_model_part_number_2]
-        )
-
-      device_1 =
-        device_fixture(realm,
-          device_id: "7mcE8JeZQkSzjLyYuh5N9A",
-          part_number: system_model_part_number_1
-        )
-
-      _device_2 =
-        device_fixture(realm,
-          device_id: "nWwr7SZiR8CgZN_uKHsAJg",
-          part_number: system_model_part_number_2
-        )
-
-      filters = %{hardware_type_name: "HW1"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with hardware_type_handle", %{realm: realm} do
-      hardware_type_1 =
-        hardware_type_fixture(name: "HW1", handle: "hw1", part_numbers: ["AAA-BBB"])
-
-      system_model_part_number_1 = "XYZ/1234"
-
-      _system_model_1 =
-        system_model_fixture(hardware_type_1,
-          name: "Foo",
-          handle: "foo",
-          part_numbers: [system_model_part_number_1]
-        )
-
-      hardware_type_2 =
-        hardware_type_fixture(name: "HW2", handle: "hw2", part_numbers: ["CCC-DDD"])
-
-      system_model_part_number_2 = "ABC/0987"
-
-      _system_model_2 =
-        system_model_fixture(hardware_type_2,
-          name: "Bar",
-          handle: "bar",
-          part_numbers: [system_model_part_number_2]
-        )
-
-      device_1 =
-        device_fixture(realm,
-          device_id: "7mcE8JeZQkSzjLyYuh5N9A",
-          part_number: system_model_part_number_1
-        )
-
-      _device_2 =
-        device_fixture(realm,
-          device_id: "nWwr7SZiR8CgZN_uKHsAJg",
-          part_number: system_model_part_number_2
-        )
-
-      filters = %{hardware_type_name: "1"}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 filters with tag", %{realm: realm} do
-      device_1 = device_fixture(realm, device_id: "7mcE8JeZQkSzjLyYuh5N9A")
-      update_attrs_1 = %{tags: ["custom", "customer"]}
-      assert {:ok, %Device{} = device_1} = Astarte.update_device(device_1, update_attrs_1)
-
-      device_2 = device_fixture(realm, device_id: "nWwr7SZiR8CgZN_uKHsAJg")
-      update_attrs_2 = %{tags: ["other"]}
-      assert {:ok, _device2} = Astarte.update_device(device_2, update_attrs_2)
-
-      filters = %{tag: "custom"}
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 combines filters with AND", %{realm: realm} do
-      device_1 = device_fixture(realm, device_id: "7mcE8JeZQkSzjLyYuh5N9A", online: true)
-      _device_2 = device_fixture(realm, device_id: "nWwr7SZiR8CgZN_uKHsAJg", online: false)
-      _device_3 = device_fixture(realm, device_id: "fsMoT420Ri-zXLjxXK6pEg", online: true)
-      filters = %{device_id: "7", online: true}
-
-      assert Astarte.list_devices(filters) == [device_1]
-    end
-
-    test "list_devices/1 returns empty list for system model filters if the device does not have a system model",
-         %{realm: realm} do
-      _device = device_fixture(realm, device_id: "7mcE8JeZQkSzjLyYuh5N9A", online: true)
-      filters = %{system_model_name: "foo"}
-
-      assert Astarte.list_devices(filters) == []
-    end
-
-    test "list_devices/1 returns empty list for hardware type filters if the device does not have a system model",
-         %{realm: realm} do
-      _device = device_fixture(realm, device_id: "7mcE8JeZQkSzjLyYuh5N9A", online: true)
-      filters = %{hardware_type_handle: "bar"}
-
-      assert Astarte.list_devices(filters) == []
-    end
-
     test "get_device!/1 returns the device with given id", %{realm: realm} do
-      device = device_fixture(realm)
+      device = astarte_device_fixture(realm)
       assert Astarte.get_device!(device.id) == device
     end
 
@@ -472,219 +171,8 @@ defmodule Edgehog.AstarteTest do
       assert {:error, %Ecto.Changeset{}} = Astarte.create_device(realm, @invalid_attrs)
     end
 
-    test "update_device/2 with valid data updates the device", %{realm: realm} do
-      device = device_fixture(realm)
-
-      update_attrs = %{
-        name: "some updated name",
-        tags: ["some", "tags"],
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "double", "value" => 42}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert device.name == "some updated name"
-      assert ["some", "tags"] == Enum.map(device.tags, & &1.name)
-      assert [custom_attribute] = device.custom_attributes
-
-      assert %Devices.Attribute{
-               namespace: :custom,
-               key: "some-attribute",
-               typed_value: %Ecto.JSONVariant{type: :double, value: 42.0}
-             } = custom_attribute
-    end
-
-    test "update_device/2 normalizes and deduplicates tags", %{realm: realm} do
-      device = device_fixture(realm)
-      update_attrs = %{tags: ["sTRANGE", "taGs   ", "  DUPlicate", "DUPLICATE"]}
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-
-      assert ["strange", "tags", "duplicate"] == Enum.map(device.tags, & &1.name)
-    end
-
-    test "update_device/2 removes tags", %{realm: realm} do
-      device = device_fixture(realm)
-      update_attrs = %{tags: ["some", "tags"]}
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert ["some", "tags"] == Enum.map(device.tags, & &1.name)
-
-      update_attrs = %{tags: ["new"]}
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert ["new"] == Enum.map(device.tags, & &1.name)
-    end
-
-    test "update_device/2 adds tags", %{realm: realm} do
-      device = device_fixture(realm)
-      update_attrs = %{tags: ["some", "tags"]}
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert ["some", "tags"] == Enum.map(device.tags, & &1.name)
-
-      update_attrs = %{tags: ["some", "tags", "new"]}
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert ["some", "tags", "new"] == Enum.map(device.tags, & &1.name)
-    end
-
-    test "update_device/2 returns an error for invalid tags", %{realm: realm} do
-      device = device_fixture(realm)
-      update_attrs = %{tags: "not a list"}
-      assert {:error, %Ecto.Changeset{}} = Astarte.update_device(device, update_attrs)
-    end
-
-    test "update_device/2 does not update the device_id", %{realm: realm} do
-      device = device_fixture(realm)
-      initial_device_id = device.device_id
-      update_attrs = %{device_id: "some updated device_id"}
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert device.device_id == initial_device_id
-    end
-
-    test "update_device/2 adds custom attributes", %{realm: realm} do
-      device = device_fixture(realm)
-
-      update_attrs = %{
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "double", "value" => 42}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert [custom_attribute] = device.custom_attributes
-
-      assert %Devices.Attribute{
-               namespace: :custom,
-               key: "some-attribute",
-               typed_value: %Ecto.JSONVariant{type: :double, value: 42.0}
-             } = custom_attribute
-
-      update_attrs = %{
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "double", "value" => 42}
-          },
-          %{
-            "namespace" => "custom",
-            "key" => "some-other-attribute",
-            "typed_value" => %{"type" => "string", "value" => "hello"}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert [^custom_attribute, new_attribute] = device.custom_attributes
-
-      assert %Devices.Attribute{
-               namespace: :custom,
-               key: "some-other-attribute",
-               typed_value: %Ecto.JSONVariant{type: :string, value: "hello"}
-             } = new_attribute
-    end
-
-    test "update_device/2 removes custom attributes", %{realm: realm} do
-      device = device_fixture(realm)
-
-      update_attrs = %{
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "double", "value" => 42}
-          },
-          %{
-            "namespace" => "custom",
-            "key" => "some-other-attribute",
-            "typed_value" => %{"type" => "string", "value" => "hello"}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert [attribute_1, _attribute_2] = device.custom_attributes
-
-      update_attrs = %{
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "double", "value" => 42}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert [^attribute_1] = device.custom_attributes
-    end
-
-    test "update_device/2 updates custom attributes", %{realm: realm} do
-      device = device_fixture(realm)
-
-      update_attrs = %{
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "double", "value" => 42}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-      assert [custom_attribute] = device.custom_attributes
-
-      assert %Devices.Attribute{
-               namespace: :custom,
-               key: "some-attribute",
-               typed_value: %Ecto.JSONVariant{type: :double, value: 42.0}
-             } = custom_attribute
-
-      update_attrs = %{
-        custom_attributes: [
-          %{
-            "namespace" => "custom",
-            "key" => "some-attribute",
-            "typed_value" => %{"type" => "string", "value" => "new value"}
-          }
-        ]
-      }
-
-      assert {:ok, %Device{} = device} = Astarte.update_device(device, update_attrs)
-
-      assert [updated_attribute] = device.custom_attributes
-
-      assert %Devices.Attribute{
-               namespace: :custom,
-               key: "some-attribute",
-               typed_value: %Ecto.JSONVariant{type: :string, value: "new value"}
-             } = updated_attribute
-    end
-
-    test "update_device/2 with invalid data returns error changeset", %{realm: realm} do
-      device = device_fixture(realm)
-      assert {:error, %Ecto.Changeset{}} = Astarte.update_device(device, @invalid_attrs)
-      assert device == Astarte.get_device!(device.id)
-    end
-
-    test "delete_device/1 deletes the device", %{realm: realm} do
-      device = device_fixture(realm)
-      assert {:ok, %Device{}} = Astarte.delete_device(device)
-      assert_raise Ecto.NoResultsError, fn -> Astarte.get_device!(device.id) end
-    end
-
     test "change_device/1 returns a device changeset", %{realm: realm} do
-      device = device_fixture(realm)
+      device = astarte_device_fixture(realm)
       assert %Ecto.Changeset{} = Astarte.change_device(device)
     end
 
@@ -695,13 +183,13 @@ defmodule Edgehog.AstarteTest do
     end
 
     test "ensure_device_exists/1 does not create a device if already existent", %{realm: realm} do
-      device = device_fixture(realm)
+      device = astarte_device_fixture(realm)
       {:ok, same_device} = Astarte.ensure_device_exists(realm, device.device_id)
       assert same_device.id == device.id
     end
 
     test "process_device_event/4 ignores unknown event", %{realm: realm} do
-      device = device_fixture(realm)
+      device = astarte_device_fixture(realm)
       device_id = device.device_id
       event = %{"type" => "unknown"}
       timestamp = DateTime.utc_now() |> DateTime.to_iso8601()
@@ -713,7 +201,7 @@ defmodule Edgehog.AstarteTest do
     test "process_device_event/4 updates online and last_connection on device_connected event", %{
       realm: realm
     } do
-      device = device_fixture(realm, %{online: false, last_connection: nil})
+      device = astarte_device_fixture(realm, %{online: false, last_connection: nil})
       assert device.online == false
       assert device.last_connection == nil
 
@@ -730,7 +218,7 @@ defmodule Edgehog.AstarteTest do
 
     test "process_device_event/4 updates online and last_disconnection on device_disconnected event",
          %{realm: realm} do
-      device = device_fixture(realm, %{online: true, last_disconnection: nil})
+      device = astarte_device_fixture(realm, %{online: true, last_disconnection: nil})
 
       assert device.online == true
       assert device.last_disconnection == nil
@@ -750,7 +238,7 @@ defmodule Edgehog.AstarteTest do
 
     test "process_device_event/4 updates serial number on incoming_data event",
          %{realm: realm} do
-      device = device_fixture(realm)
+      device = astarte_device_fixture(realm)
 
       assert device.serial_number == nil
 
@@ -773,12 +261,12 @@ defmodule Edgehog.AstarteTest do
 
     test "process_device_event/4 updates part number on incoming_data event",
          %{realm: realm} do
-      device = device_fixture(realm)
+      device = astarte_device_fixture(realm)
       assert device.part_number == nil
 
       part_number = "XYZ123"
 
-      system_model = system_model_fixture(hardware_type_fixture(), part_numbers: [part_number])
+      _system_model = system_model_fixture(hardware_type_fixture(), part_numbers: [part_number])
 
       device_id = device.device_id
 
@@ -793,122 +281,9 @@ defmodule Edgehog.AstarteTest do
       timestamp_string = DateTime.to_iso8601(timestamp)
       assert :ok = Astarte.process_device_event(realm, device_id, event, timestamp_string)
 
-      assert device =
-               device.id
-               |> Astarte.get_device!()
-               |> Astarte.preload_system_model_for_device()
+      device = Astarte.get_device!(device.id)
 
       assert device.part_number == part_number
-      assert device.system_model.id == system_model.id
-    end
-
-    test "get_device_capabilities/1 returns all capabilities if interfaces are implemented by the device" do
-      device_introspection = %{
-        "io.edgehog.devicemanager.BaseImage" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.BatteryStatus" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.CellularConnectionProperties" => %InterfaceVersion{
-          major: 0,
-          minor: 1
-        },
-        "io.edgehog.devicemanager.CellularConnectionStatus" => %InterfaceVersion{
-          major: 0,
-          minor: 1
-        },
-        "io.edgehog.devicemanager.Commands" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.HardwareInfo" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.LedBehavior" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.NetworkInterfaceProperties" => %InterfaceVersion{
-          major: 0,
-          minor: 1
-        },
-        "io.edgehog.devicemanager.OSInfo" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.RuntimeInfo" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.OTARequest" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.OTAResponse" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.StorageUsage" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.SystemInfo" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.SystemStatus" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.config.Telemetry" => %InterfaceVersion{major: 0, minor: 1},
-        "io.edgehog.devicemanager.WiFiScanResults" => %InterfaceVersion{major: 0, minor: 1}
-      }
-
-      expected_capabilities = [
-        :base_image,
-        :battery_status,
-        :cellular_connection,
-        :commands,
-        :geolocation,
-        :hardware_info,
-        :led_behaviors,
-        :network_interface_info,
-        :operating_system,
-        :runtime_info,
-        :software_updates,
-        :storage,
-        :system_info,
-        :system_status,
-        :telemetry_config,
-        :wifi
-      ]
-
-      assert Enum.sort(expected_capabilities) ==
-               Enum.sort(Astarte.get_device_capabilities(device_introspection))
-    end
-
-    test "get_device_capabilities/1 returns a capability only if all its interfaces are supported by the device" do
-      partial_introspection_1 = %{
-        "io.edgehog.devicemanager.OTARequest" => %InterfaceVersion{major: 0, minor: 1}
-      }
-
-      partial_introspection_2 = %{
-        "io.edgehog.devicemanager.OTAResponse" => %InterfaceVersion{major: 0, minor: 1}
-      }
-
-      assert :software_updates not in Astarte.get_device_capabilities(partial_introspection_1)
-      assert :software_updates not in Astarte.get_device_capabilities(partial_introspection_2)
-    end
-
-    test "get_device_capabilities/1 returns only geolocation if no interface is supported by the device" do
-      assert [:geolocation] = Astarte.get_device_capabilities(%{})
-    end
-
-    test "get_device_capabilities/1 should not fail when devices uses a minor greater than the one required" do
-      device_introspection = %{
-        "io.edgehog.devicemanager.BatteryStatus" => %InterfaceVersion{major: 0, minor: 2},
-        "io.edgehog.devicemanager.CellularConnectionProperties" => %InterfaceVersion{
-          major: 0,
-          minor: 1
-        },
-        "io.edgehog.devicemanager.CellularConnectionStatus" => %InterfaceVersion{
-          major: 0,
-          minor: 3
-        }
-      }
-
-      expected_capabilities = [
-        :battery_status,
-        :cellular_connection,
-        :geolocation
-      ]
-
-      assert Enum.sort(expected_capabilities) ==
-               Enum.sort(Astarte.get_device_capabilities(device_introspection))
-    end
-
-    test "get_device_capabilities/1 should not return a capability if major version mismatches" do
-      device_introspection = %{
-        "io.edgehog.devicemanager.BatteryStatus" => %InterfaceVersion{major: 1, minor: 0},
-        "io.edgehog.devicemanager.CellularConnectionProperties" => %InterfaceVersion{
-          major: 1,
-          minor: 1
-        },
-        "io.edgehog.devicemanager.CellularConnectionStatus" => %InterfaceVersion{
-          major: 0,
-          minor: 1
-        }
-      }
-
-      assert [:geolocation] = Astarte.get_device_capabilities(device_introspection)
     end
   end
 end
