+++ backend/test/edgehog/devices/selector_test.exs
@@ -22,7 +22,8 @@ defmodule Edgehog.Devices.SelectorTest do
   use Edgehog.DataCase
 
   import Edgehog.AstarteFixtures
-  alias Edgehog.Astarte
+  import Edgehog.DevicesFixtures
+  alias Edgehog.Devices
   alias Edgehog.Devices.Selector
   alias Edgehog.Devices.Selector.AST.AttributeFilter
   alias Edgehog.Devices.Selector.AST.BinaryOp
@@ -337,18 +338,18 @@ defmodule Edgehog.Devices.SelectorTest do
     end
   end
 
-  defp add_tags(%Astarte.Device{} = device, tags) do
-    {:ok, device} = Astarte.update_device(device, %{tags: tags})
+  defp add_tags(%Devices.Device{} = device, tags) do
+    {:ok, device} = Devices.update_device(device, %{tags: tags})
     device
   end
 
-  defp add_custom_attributes(%Astarte.Device{} = device, kv_map) do
+  defp add_custom_attributes(%Devices.Device{} = device, kv_map) do
     custom_attributes =
       for {k, v} <- kv_map, k = to_string(k) do
         %{namespace: :custom, key: k, typed_value: v}
       end
 
-    {:ok, device} = Astarte.update_device(device, %{custom_attributes: custom_attributes})
+    {:ok, device} = Devices.update_device(device, %{custom_attributes: custom_attributes})
     device
   end
 end
